refactor(acp): flatten session/adapters/ into session/
InputPreprocessor and McpConfig are not adapters — they are plain
data-transform utilities consumed only by AcpSession. Remove the
unnecessary nesting.
